backend/requirements.txt
# FastAPI et serveur ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données et ORM
sqlmodel==0.0.14
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Variables d'environnement
python-dotenv==1.0.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS
python-cors==1.7.0

# Validation de données
pydantic==2.5.0
email-validator==2.1.0

# Dates et temps
python-dateutil==2.8.2

# Client Supabase (optionnel)
supabase==2.1.0